implemented fh as kue job
