+++ views/player_index.jade
@@ -3,28 +3,27 @@ extend player
 block playercontent
   if (user && user.account_id === player.account_id)
     .row
-      .col-md-6
+      .col-md-12
         h3 Welcome!
-        ul
-          li Just joined?  We probably don't have your old matches yet.
-          li Play a game and come back, or request a match by ID.
-          li Keep using YASP and we'll get your full match history.
-      .col-md-6
-        h3 Actions
+        if (!user.full_history_time)
+          .alert.alert-danger We don't have your old matches yet!  Check back later to see if we've gotten your full history.  In the meantime, play a game and we'll parse it for you!
+      .col-md-3
         .checkbox
           label
             input.pref#dark(type="checkbox", checked=user.dark_theme)
             | Use Dark Theme
+      .col-md-3
         .checkbox
           label
             input.pref#publicmmr(type="checkbox", checked=user.publicmmr)
             | Make MMR public
-        p: a.btn.btn-success(href="/request") Request by ID &raquo;
+      .col-md-3
         if (user.account_id in ratingPlayers)
           p: button.btn.btn-default(disabled) MMR Tracked
         else
-          p: a.btn.btn-success(href="http://steamcommunity.com/profiles/"+ (bots[0] ? bots[0].steamID : "")) Add MMR Tracker &raquo;
-
+          p: a.btn.btn-warning(href="http://steamcommunity.com/profiles/"+ (bots[0] ? bots[0].steamID : "")) Add MMR Tracker &raquo;
+      .col-md-3
+        p: a.btn.btn-primary(href="/request") Request by ID &raquo;
   .row
     .col-md-6
       h3 Recent Matches
